add --semantic flag
Like --final, --semantic compensates for differing platform conventions
regarding termination of the last line in a file. In contrast,
--semantic aims to preserve logical lines across representations.
This assumes windows conventions use CRLF as line separators, not terminators.
An equivalent formulation: lines are terminated by exactly one of CRLF or EOF.
Yet another: CRLF always starts a new line, even at file end.

The goal is to cope with the unterminated-line problem in a way that
allows lossless round-trips. In concept, git smudge and clean filters are
intended to be the inverse of one another, making changes that are fully
reversible. Adding a trailing blank line to a windows file is normally
inoffensive, but can be a nuisance in the presence of change-tracking.